type

+++ mujoco_playground/_src/manipulation/ego_sim/pick.py
@@ -205,7 +205,7 @@ class RUMPickCube(rum.RUMGripper):
         state.metrics.update(
             {
                 "reward/lifted": lifted,
-                "reward/success": success,
+                "reward/success": success.astype(float),
             }
         )
